Rename implementation classes according to their interfaces
