+++ code/core/src/main/java/org/betonquest/betonquest/api/quest/objective/service/DefaultEventServiceSubscriptionBuilder.java
@@ -37,7 +37,7 @@ public class DefaultEventServiceSubscriptionBuilder<T extends Event> implements 
     /**
      * The event service to register events with.
      */
-    private final DefaultObjectiveService eventService;
+    private final DefaultObjectiveServiceProvider eventService;
 
     /**
      * The event class to register.
@@ -95,7 +95,7 @@ public class DefaultEventServiceSubscriptionBuilder<T extends Event> implements 
      * @param eventService the event service to register events with
      * @param eventClass   the event class to register
      */
-    public DefaultEventServiceSubscriptionBuilder(final DefaultObjectiveService eventService, final Class<T> eventClass) {
+    public DefaultEventServiceSubscriptionBuilder(final DefaultObjectiveServiceProvider eventService, final Class<T> eventClass) {
         this.eventService = eventService;
         this.eventClass = eventClass;
         this.objectiveID = null;
